fix UnmarshalText

+++ internal/ethapi/transaction_args.go
@@ -51,7 +51,7 @@ type TransactionArgs struct {
 	Input *hexutil.Bytes `json:"input"`
 
 	// AltFeeTxType
-	FeeTokenID *hexutil.Uint64 `json:"feeTokenID,omitempty"`
+	FeeTokenID *hexutil.Uint16 `json:"feeTokenID,omitempty"`
 	FeeLimit   *hexutil.Big    `json:"feeLimit,omitempty"`
 
 	// Introduced by AccessListTxType transaction.
